iota_rest_kv
Module kv_store_client

This module provides a client for interacting with the key-value store.

Structs

AwsStatus

Status of the AWS components used by the KvStoreClient.

CachedAwsStatus

Represents a cached status of AWS components.

DynamoDbConfig

Configuration for DynamoDB connection.

KvStoreClient

Provides read access to data ingested by the iota-data-ingestion crate’s KVStoreWorker.

KvStoreConfig

Configuration for the KvStoreClient used to access data from S3 and DynamoDB.

ServiceStatus

Represents the health status of a service.

Constants

AWS_STATUS_CACHE_TTL 🔒

CONNECT_TIMEOUT_SECS 🔒

OPERATION_ATTEMPT_TIMEOUT_SECS 🔒

OPERATION_TIMEOUT_SECS 🔒