iota_rest_kv::server

Function fallback

async fn fallback() -> impl IntoResponse

Handles requests to routes that are not defined in the API.

This fallback handler is called when the requested URL path does not match any of the defined routes. It returns a 404 Not Found error, indicating that the requested resource could not be found. This can happen if the user enters an incorrect URL or if the requested resource (identified by a Key) cannot be extracted from the request.