Crate iota_rest_kv

Modules

errors 🔒

This module includes the error types the REST API sends back to the client.

extractors 🔒

This module includes custom extractors needed for validation and custom errors messages on the input provided by the client.

kv_store_client 🔒

This module provides a client for interacting with the key-value store.

routes 🔒

This module contains all routes supported by the REST API server

server 🔒

This module includes helper wrappers for building and starting a REST API server.

types 🔒

This module includes types and useful conversions.

Structs

Cli 🔒

The main CLI application.

RestApiConfig

Functions

init_tracing 🔒

Initialize the tracing with custom subscribers.

main 🔒

shutdown_signal_listener 🔒

Set up a CTRL+C & SIGTERM handler for graceful shutdown and spawn a tokio task.